// types.ts
export enum IssueCategory {
  Litter = 'Litter',
  OverflowingBin = 'Overflowing Bin',
  ConstructionDebris = 'Construction Debris',
  BlockedDrainage = 'Blocked Drainage',
  Graffiti = 'Graffiti',
  BrokenBench = 'Broken Public Bench',
}

export interface Issue {
  id: number;
  user: string;
  userAvatar: string;
  category: IssueCategory;
  imageUrl: string;
  afterImageUrl?: string;
  location: string;
  ward: string;
  description: string;
  timestamp: string;
  status: 'Reported' | 'In Progress' | 'Solved';
  volunteersNeeded?: number;
  participants?: string[];
  coordinates: { x: number; y: number; };
  eventDetails?: {
      date: string;
      time: string;
  };
}

export interface UserStats {
    reportsMade: number;
    eventsOrganized: number;
    eventsJoined: number;
    quizCompleted: boolean;
    recyclingLogs: number;
    supplyKitsPickedUp: number;
    supplyKitsToday: number;
    microActionsLogged: number;
    microActionsToday: number;
    disturbanceReports: number;
    safetyKitsRedeemed: number;
}

export interface Activity {
  id: string;
  user: {
    name: string;
    avatar: string;
    isAdmin?: boolean;
  };
  type: 'reported' | 'organized' | 'joined' | 'redeemed' | 'quiz_completed' | 'forum_thread_created' | 'forum_reply' | 'recycled_item' | 'supply_kit_pickup' | 'micro_action_logged' | 'announcement' | 'safety_kit_redeemed' | 'merchandise_purchased' | 'merchandise_review';
  description: string;
  timestamp: number; // Changed to number for sorting
  pointsChange: number;
  issueId?: number;
  isAnnouncement?: boolean;
}

export interface PurchaseReceipt {
  id: number;
  userName: string;
  rewardTitle: string;
  rewardPartner: string;
  rewardImageUrl: string;
  costInSP: number;
  pointsUsed: number;
  amountPaidNPR: number;
  timestamp: number;
  qrCodeUrl: string;
  status: 'pending' | 'confirmed';
}

export interface AdminPurchaseReceipt extends PurchaseReceipt {
  userAvatar: string;
}

export interface UserRank {
  rank: number;
  name: string;
  avatar: string;
  points: number;
  stats: UserStats;
  ward: string;
  activity: Activity[];
  purchaseHistory: PurchaseReceipt[];
  isAdmin?: boolean;
  notifications?: string[];
}